Second pass in separation of source from header files

+++ examples/libstreaming/streaming/stream_consumer.h
@@ -25,12 +25,12 @@ class stream_consumer : public rpc::rpc_client {
     }
     int64_t& buf_off = read_buffer_.first;
     std::string& buf = read_buffer_.second;
-    int64_t rbuf_lim = buf_off + buf.size();
+    int64_t rbuf_lim = static_cast<int64_t>(buf_off + buf.size());
     if (buf_off == -1 || offset_ < buf_off || offset_ >= rbuf_lim) {
       read_buffer_.first = offset_;
       client_->read(buf, cur_multilog_id_, buf_off, prefetch_size_);
     }
-    _return = buf.substr(offset_ - buf_off, cur_schema_.record_size());
+    _return = buf.substr(static_cast<unsigned long>(offset_ - buf_off), cur_schema_.record_size());
     offset_ += cur_schema_.record_size();
   }
